refactor(mcp): split get_context http tool tests
